fix : columns based on survey form type
