+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2023-11-20T15:41:02.046Z\n"
-"PO-Revision-Date: 2023-11-20T15:41:02.046Z\n"
+"POT-Creation-Date: 2023-11-22T06:15:51.058Z\n"
+"PO-Revision-Date: 2023-11-22T06:15:51.058Z\n"
 
 msgid "WHO privacy policy"
 msgstr ""
@@ -23,22 +23,22 @@ msgstr ""
 msgid "Save"
 msgstr ""
 
-msgid "PPS Survey Forms"
+msgid "PPS Surveys"
 msgstr ""
 
 msgid "New Survey"
 msgstr ""
 
-msgid "PPS Country Questionnaires"
+msgid "Country Questionnaires"
 msgstr ""
 
-msgid "PPS Hospital Forms"
+msgid "Hospital Forms"
 msgstr ""
 
-msgid "PPS Ward Registers"
+msgid "Ward Registers"
 msgstr ""
 
-msgid "PPS Patient Registers"
+msgid "Patient Registers"
 msgstr ""
 
 msgid "Submission Success!"
@@ -47,9 +47,6 @@ msgstr ""
 msgid ""
 "Submission Failed! You do not have the necessary permissions, please "
 "contact your administrator"
-msgstr ""
-
-msgid "Create New Survey"
 msgstr ""
 
 msgid "PPS Survey Name"
@@ -68,6 +65,9 @@ msgid "Assigned Org Unit"
 msgstr ""
 
 msgid "Action"
+msgstr ""
+
+msgid "PPS Survey Forms"
 msgstr ""
 
 msgid "Yes"
